Optimise method of spliting up arguments and options in a command
This fixes bug with print command where print and an option specified would cause an exception. Note the addition of fullcommand to Command.java, this now stores the whole original command. ExtractOptions is now called addOptions and is called with string of options, from splitargs, it splits the arguments itself and adds options to the given command